feat : 가게 전체리뷰 조회

+++ backend/main/src/main/java/com/ssafy/c107/main/domain/members/dto/response/SellerReviewAllResponse.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @Builder
 public class SellerReviewAllResponse {
 
-    private int storeId;
+    private Long storeId;
     private int positiveCnt;
     private int negativeCnt;
 }
